Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/dom/nodes/rootNode-expected.txt


PASS getRootNode() must return context object's shadow-including root if options's composed is true, and context object's root otherwise
PASS getRootNode() must return the context object when it does not have any parent
PASS getRootNode() must return the parent node of the context object when the context object has a single ancestor not in a document
PASS getRootNode() must return the document when a node is in document
PASS getRootNode() must return a document fragment when a node is in the fragment
